requirements.txt
influxdb-client==1.35.0
pygazpar==1.2.0
python-dotenv==0.21.0
